Adding azure-mgmt-core to requirements.txt
Adding exceptions to test_pkg_imports.py to skip vt and vt_graph_api (now extras)

+++ tests/data/drivers/test_mordor_driver.py
@@ -40,7 +40,8 @@ def qry_provider():
         Path(_SAVE_PATH).unlink()
     for file in Path(_SAVE_FOLDER).glob("*"):
         Path(file).unlink()
-    Path(_SAVE_FOLDER).rmdir()
+    if Path(_SAVE_FOLDER).is_dir():
+        Path(_SAVE_FOLDER).rmdir()
 
 
 # pylint: disable=redefined-outer-name, protected-access, global-statement
